requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
httpx==0.25.2
email-validator==2.1.0
python-multipart==0.0.6